Removes AttributedString/LayoutManager
