+++ lib/android/modulePaths.js
@@ -121,13 +121,8 @@ module.exports = [
   "node_modules/react/cjs/react.development.js",
   "node_modules/react/index.js",
   "packages/pages/src/pages.js",
-  "packages/typeset/src/AttributedString.ts",
   "packages/typeset/src/BoxExclusion.ts",
   "packages/typeset/src/FontStorage.ts",
-  "packages/typeset/src/LayoutManager.ts",
   "packages/typeset/src/Point.ts",
-  "packages/typeset/src/PositionedItem.ts",
-  "packages/typeset/src/Span.ts",
-  "packages/typeset/src/TextContainer.ts",
-  "packages/typeset/src/index.ts"
+  "packages/typeset/src/index.ts",
 ];
