--- packages/typeset/__tests__/PositionedItem.test.ts
@@ -1,8 +1,0 @@
-import PositionedItem from "../src/PositionedItem";
-
-test("PositionedItem#constructor", () => {
-  expect(() => {
-    const newItem = new PositionedItem({} as any, {} as any);
-    return newItem;
-  }).not.toThrow();
-});
