+++ packages/typeset/index.ts
@@ -1,8 +1,3 @@
-export { default as LayoutManager } from "./src/LayoutManager";
 export { default as BoxExclusion } from "./src/BoxExclusion";
-export { default as TextContainer } from "./src/TextContainer";
-export { default as AttributedString } from "./src/AttributedString";
-export { default as PositionedItem } from "./src/PositionedItem";
-export * from "./src/AttributedString";
 export * from "./src/BoxExclusion";
 export { default as FontStorage } from "./src/FontStorage";
