--- packages/typeset/src/PositionedItem.ts
@@ -1,12 +1,0 @@
-import AttributedString from "./AttributedString";
-import Point from "./Point";
-
-export default class PositionedItem {
-  public text: AttributedString;
-  public position: Point;
-
-  constructor(text: AttributedString, position: Point) {
-    this.text = text;
-    this.position = position;
-  }
-}
