+++ packages/typeset/src/index.ts
@@ -1,8 +1,3 @@
-export { default as LayoutManager } from "./LayoutManager";
 export { default as BoxExclusion } from "./BoxExclusion";
-export { default as TextContainer } from "./TextContainer";
-export { default as AttributedString } from "./AttributedString";
-export { default as PositionedItem } from "./PositionedItem";
-export * from "./AttributedString";
 export * from "./BoxExclusion";
-export { default as FontStorage } from "./FontStorage";
+export { default as FontStorage, TypographySettings } from "./FontStorage";
